Codacy improvements
