+++ api/python/provisioner/commands/set_cluster_id.py
@@ -25,32 +25,16 @@ from .. import inputs
 from ..api import grains_get
 from ..config import LOCAL_MINION
 
-from provisioner.commands import (
-    PillarSet
-)
 from . import CommandParserFillerMixin
+from . import (
+    PillarSet, GetClusterId
+)
+from .grains_get import GrainsGet
 
 from ..vendor import attr
 from provisioner.salt import local_minion_id
 
 logger = logging.getLogger(__name__)
-
-# TODO TEST
-@attr.s(auto_attribs=True)
-class GetClusterId(CommandParserFillerMixin):
-    input_type: Type[inputs.NoParams] = inputs.NoParams
-    _run_args_type = RunArgsEmpty
-
-    def run(self):
-        # return list(function_run(
-        #     'grains.get',
-        #     fun_args=['cluster_id']
-        # ).values())[0]
-
-        PillarGet().run(
-            'cluster/cluster_id',
-            targets=local_minion_id()
-        )[local_minion_id()]
 
 
 @attr.s(auto_attribs=True)
@@ -89,46 +73,25 @@ class SetClusterId(CommandParserFillerMixin):
                 "Setting Cluster Id on Primary node of the cluster."
             )
 
-            cluster_id_from_grains = grains_get(
-                "cluster_id",
-                local_minion_id()
-            )[local_minion_id()]["cluster_id"]
+            cluster_id = cluster_id if cluster_id else GetClusterId().run()
 
-            cluster_id_from_pillar = PillarGet().run(
-                'cluster/cluster_id',
-                targets=local_minion_id()
-            )[local_minion_id()]
-            cluster_id_from_pillar = GetClusterId.run()
-
-            # Check if Cluster ID is set in Grains
-
-            if not cluster_id_from_grains:
+            if not cluster_id:
                 logger.info(
-                    "ClusterID is not found in grains data. Generating one.."
+                    "Existing ClusterID is neither set "
+                    "nor is provided as user input. "
+                    "Generating a new cluster_id"
                 )
                 cluster_uuid = str(uuid.uuid4())
                 logger.info("Setting the generated ClusterID across all nodes..")
 
-                PillarSet().run(
-                    'cluster/cluster_id',
-                    f'{cluster_uuid}',
-                    targets=local_minion_id()
-                )
-
-            elif cluster_id_from_grains and not cluster_id_from_pillar:
-                logger.info(
-                    "ClusterID is not set in pillar data. Proceeding to set now.."
-                )
-                PillarSet().run(
-                    'cluster/cluster_id',
-                    f'{cluster_id_from_grains}',
-                    targets=local_minion_id()
-                )
-
-            else:
-                logger.info(
-                    "Bootstrapping completed and ClusterID is already set!"
-                )
+            logger.info(
+                "Set cluster_id in pillar cluster."
+            )
+            PillarSet().run(
+                'cluster/cluster_id',
+                f'{cluster_uuid}',
+                targets=local_minion_id()
+            )
 
             logger.info(
                 "Success: Unique ClusterID assignment to pillar data."
